🐛 [-bug] Ensure that the interface for BuildExeBuildInfo matches
Build files that use BuildExeBuildInfo may be invoked in environments that aren't based on Common_PythonDevelopment. In those cases, the build file should be able to seamlessly transition to BuildInfoBase in Common_Foundation.BuildImpl (which will mostly likely display a message saying that the build is not available in this environment configuration).

These changes ensure that the parameters taken in BuildExeBuildInfo's constructor are a subset of those taken in BuildInfoBase's constructor.