🎨 [refactor] Updated BuildPythonExecutable code
The previous iteration of this code implemented this functionality as a base class. Unfortunately this didn't work very well because the repositories that used the functionality invoked it from different environment configurations (some of which didn't have access to Common_PythonDevelopment). This would result in errors associated with invalid imports in those configurations without a dependency on Common_PythonDevelopment.

This new version implements the functionality in functions (rather than as a base class), giving the caller more freedom in how they invoke it within different environment configurations.